# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YouCare Swagger Documentation",
        "description": "Implementation of Swagger in Laravel REST API",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "My API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user and generate JWT token",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competences",
                        "in": "query",
                        "description": "User's competences",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/listing": {
            "get": {
                "tags": [
                    "Listing"
                ],
                "summary": "Fetch all listings",
                "operationId": "listing-all",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listings fetched successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Listing"
                ],
                "summary": "Create a new listing",
                "operationId": "listing-add",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Listing's title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Listing's description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Listing's date",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Listing's location",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competences",
                        "in": "query",
                        "description": "Listing's competences",
                        "required": true,
                        "schema": {
                            "type": "json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing added successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/listing/{listing}": {
            "put": {
                "tags": [
                    "Listing"
                ],
                "summary": "Update a new listing",
                "operationId": "listing-update",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Listing's title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Listing's description",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Listing's date",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Listing's location",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competences",
                        "in": "query",
                        "description": "Listing's competences",
                        "schema": {
                            "type": "json"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listing updated successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/apply": {
            "put": {
                "tags": [
                    "Application"
                ],
                "summary": "Add a new application",
                "operationId": "application-add",
                "parameters": [
                    {
                        "name": "listing_id",
                        "in": "query",
                        "description": "Listing's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application added successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/approve": {
            "put": {
                "tags": [
                    "Application"
                ],
                "summary": "Approve an application",
                "operationId": "application-approve",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "query",
                        "description": "Application's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application approved successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/decline": {
            "put": {
                "tags": [
                    "Application"
                ],
                "summary": "Decline an application",
                "operationId": "application-decline",
                "parameters": [
                    {
                        "name": "application_id",
                        "in": "query",
                        "description": "Application's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application declined successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based auth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}